Change display of Project and Health views to test with users.
Based on feedback we learned that it was not obvious how to manage project
health. The layout for both the project and health pages has been changed.

Some style changes were made to remove the circle elements.

As part of this work unused files were deleted to make it easier to reason
about what is needed as part of the product, it took me some time to figure out
whether the deleted files actually contained useful code.

There was also code removed from the Project page that we don't use, for example
specific service manager and open source display options.