Improve consistency of layout across pages
Different styles were used on pages, as well as different column layouts.
It meant the main elements of the page 'jump' around which is a bit
inconsistent, and results in the user having to adjust focus frequently.

Improving layout consistency helps with this.

Refactored some code as part of this change to remove unnecessary files.