Improve layout and styling to make health more obvious
Based on user feedback we have made a number of changes to the layout
and styling of the dashboard to make data more accessible and obvious.

Corrected a number of incorrectly used classes on headers so proper
spacing is in place around header elements, new health block style to
allow a more 'radiator' style display and removed some content and text
that was redundant to reduce clutter.