To improve visibility of risk on projects adding a 'test' risk radar
The risk radar concept is in use on a few projects out in the wild,
it shows risks categorised by proximity and 'area', teams have found
this useful to visual what is important to track and resolve. This
commit adds a proof of concept 'test' version that mimics the current
whiteboard based approach.

It uses an HTML5 canvas and some simple shapes to create the radar
layout and some 'mock' risks have been added.