Moving to using docker for running locally and for deployments
Dockerising the dashboard so we can experiment with running something
real using a container based approach. Using Kubernetes to host locally.

Dockerfile creates the image, which is published to our public repo on
dockerhub.

Kubernetes deployment yaml file is used to declaritvely state how we
want to run it on a Kubernetes cluster.