Removed obvious styling for health status
The rather obvious styling for projects made it difficult to use on
customer site, especially if we considered the project to require
additional support (i.e. risky).

It has been changed now to be a text based indicator that also describes
the project need rather than the simple status.

In order to achieve this a mapping function was added to the project model,
this isn't perfect, and should be extracted to a better location later, but for
now the important change is the user facing one.